// components/Contact/index.tsx
"use client";

import { useForm } from "react-hook-form";
import { zodResolver } from "@hookform/resolvers/zod";

import {
  contactSchema,
  ContactFormValues
} from "@/lib/contact-schema";
import { sendEmail } from "@/actions/send-email";

import {
  Form,
  FormControl,
  FormField,
  FormItem,
  FormLabel,
  FormMessage,
} from "@/components/ui/form";

import { Input } from "@/components/ui/input";
import { Textarea } from "@/components/ui/textarea";
import { Button } from "@/components/ui/button";
import { toast } from "sonner";

export default function ContactForm() {
  const form = useForm<ContactFormValues>({
    resolver: zodResolver(contactSchema),
    mode: "onChange",
    defaultValues: {
      name: "",
      email: "",
      message: "",
    },
  });

  const {
    handleSubmit,
    formState: { isSubmitting },
    setError,
    reset,
  } = form;

  async function onSubmit(values: ContactFormValues) {
    const result = await sendEmail(values);

    if (!result.success) {
      if (result.fieldErrors) {
        Object.entries(result.fieldErrors).forEach(([field, messages]) => {
          setError(field as keyof ContactFormValues, {
            message: Array.isArray(messages) ? messages[0] : messages,
          });
        });
      } else {
        toast.error(
          result.message || "Something went wrong. Please try again later.",
        );
      }
      return;
    } else {
      reset();
      toast.success(
        result.message || "Your message has been sent successfully!",
      );
    }
  }

  return (
    <section className="max-w-xl space-y-6">
      <Form {...form}>
        <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
          <FormField
            control={form.control}
            name="name"
            render={({ field }) => (
              <FormItem className="h-20">
                <FormLabel>Name</FormLabel>
                <FormControl>
                  <Input placeholder="Your name" {...field} />
                </FormControl>
                <FormMessage />
              </FormItem>
            )}
          />

          <FormField
            control={form.control}
            name="email"
            render={({ field }) => (
              <FormItem className="h-20">
                <FormLabel>Email</FormLabel>
                <FormControl>
                  <Input placeholder="Your email" {...field} />
                </FormControl>
                <FormMessage />
              </FormItem>
            )}
          />

          <FormField
            control={form.control}
            name="message"
            render={({ field }) => (
              <FormItem className="h-20">
                <FormLabel>Message</FormLabel>
                <FormControl className="h-30">
                  <Textarea
                    placeholder="Your message here..."
                    rows={10}
                    {...field}
                  />
                </FormControl>
                <FormMessage />
              </FormItem>
            )}
          />

          <Button
            type="submit"
            disabled={isSubmitting}
            className="border px-6 py-4 mt-22"
          >
            {isSubmitting ? "Sending..." : "Say Hello"}
          </Button>
        </form>
      </Form>
    </section>
  );
}
